A faster implementation of `GetCommitsInfo`, particularly for repositories with many files. Addresses https://github.com/go-gitea/gitea/issues/491 and https://github.com/go-gitea/gitea/issues/502.

A revival of the failed #53 (which was reverted by #73).

__BENCHMARKS__

Old implementations:
```
BenchmarkEntries_GetCommitsInfo/gitea-4         	     200	  89001816 ns/op
BenchmarkEntries_GetCommitsInfo/manyfiles-4     	       1	47869780305 ns/op
BenchmarkEntries_GetCommitsInfo/moby-4          	     100	 174120584 ns/op
BenchmarkEntries_GetCommitsInfo/go-4            	      20	 727309260 ns/op
BenchmarkEntries_GetCommitsInfo/linux-4         	      20	 853282140 ns/op
```

New implementation:
```
BenchmarkEntries_GetCommitsInfo/gitea-4         	     200	  79968242 ns/op
BenchmarkEntries_GetCommitsInfo/manyfiles-4     	      50	 261940680 ns/op
BenchmarkEntries_GetCommitsInfo/moby-4          	     100	 154584374 ns/op
BenchmarkEntries_GetCommitsInfo/go-4            	      20	 715422795 ns/op
BenchmarkEntries_GetCommitsInfo/linux-4         	      20	 833090406 ns/op
```

In summary, we see a >150x improvement in repos with many files, and mild improvements in other repos.

__IMPLEMENTATION DETAILS__:
Runs targeted and untargeted searches in parallel. The untargeted search (1 thread) sifts through the output of a `git log --name-status -c -- <tree-path>` command, seeing which files the most recent commits have affected. The targeted search (multiple threads) runs `git rev-list -1 HEAD -- <entry>` commands to get the latest commit for specific file that have not been found by the untargeted search. The targeted search is similar to the old implementation, with a few enhancements (e.g. uses `git rev-list` instead of `git log`).

If there are only a few remaining entries, then the untargeted search stops to allow the targeted-search threads to finish without any interference (in my tests, this interference had a non-negligible impact on performance).

__TESTING__:
I have not been able to test this new implementation on a Windows machine. If someone could test the new implementation on Windows (to check for forward-slash vs. backslash bugs), that would be great.