@lafriks 2017-12-09
@ethantkoenig Is it really worth risking locking up if unexpected error occurs? Can you retest test how big performance penalty in this case it is with `defer`?