@ethantkoenig 2017-12-10
@lafriks Overhead of using `defer` was less than I expected (essentially neglible), so I've switched to using `defer`.